edit files according to stuff notes at the course site

+++ hw2.c
@@ -2,10 +2,10 @@
 #include <linux/sched.h>
 #include <linux/kernel.h>
 
-int search_children(struct task_struct* task) {
+long search_children(struct task_struct* task) {
 	struct task_struct *child;
 	struct list_head *list;
-	int sum = task->weight;
+	long sum = task->weight;
 	//printk("Now in search children with PID %d\n", task->pid);
 	//printk("The weight is: %d\n", task->weight);
 	list_for_each(list, &task->children) {
@@ -22,15 +22,15 @@ asmlinkage long sys_hello(void) {
 	return 0;
 }
 
-asmlinkage int sys_get_total_weight(void) {
+asmlinkage long sys_get_total_weight(void) {
 	//int sum = get_children_weight(current);
-	int sum = search_children(current);
+	long sum = search_children(current);
 	//printk("The current pid is: %d\n", current->pid);
 	//printk("The total weight is: %d\n", sum);
 	return sum;
 }
 
-asmlinkage int sys_set_weight(int weight) {
+asmlinkage long sys_set_weight(int weight) {
 	if (weight < 0) {
 		errno = -EINVAL
 		return -1;
@@ -45,7 +45,7 @@ asmlinkage int sys_set_weight(int weight) {
 asmlinkage long sys_get_heaviest_child(void) {
 	struct task_struct *child;
         struct list_head *list;
-	int max_weight = 0;
+	long max_weight = 0;
 	pid_t max_pid = 0;
 
         list_for_each(list, &current->children) {
